refactor systemd setup
